prov/shm: Make progress errors ints instead of unit64
Make progress errors ints to match the fi_cq_err_entry
struct's err type. This error needs to be positive so
the paths have been checked and they all return negative
errors. The err sign is flipped when calling write_err_comp
to keep it positive like the API expects.

+++ prov/shm/src/smr.h
@@ -290,7 +290,7 @@ typedef ssize_t (*smr_proto_func)(struct smr_ep *ep, struct smr_region *peer_smr
 extern smr_proto_func smr_proto_ops[smr_src_max];
 
 int smr_write_err_comp(struct util_cq *cq, void *context,
-		       uint64_t flags, uint64_t tag, uint64_t err);
+		       uint64_t flags, uint64_t tag, int err);
 int smr_complete_tx(struct smr_ep *ep, void *context, uint32_t op,
 		    uint64_t flags);
 int smr_complete_rx(struct smr_ep *ep, void *context, uint32_t op,
